requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
psycopg2
psycopg2-binary
drf-spectacular